move Range to diagnostics
