+++ src/compiler.ts
@@ -15,6 +15,7 @@ import {
 } from "./builtins";
 
 import {
+  Range,
   DiagnosticCode,
   DiagnosticEmitter
 } from "./diagnostics";
@@ -108,7 +109,6 @@ import {
 
 import {
   Token,
-  Range,
   operatorTokenToString
 } from "./tokenizer";
